ergasia-api/requirements.txt
alembic==1.5.4
asyncpg==0.22.0
bcrypt==3.2.0
cffi==1.14.5
chardet==4.0.0
click==7.1.2
cryptography==3.4.6
databases==0.4.1
docx-mailmerge==0.5.0
ecdsa==0.14.1
et-xmlfile==1.1.0
fastapi-sqlalchemy==0.2.1
h11==0.12.0
idna==2.10
lxml==4.6.3
mako==1.1.4
markupsafe==1.1.1
numpy==1.20.3
openpyxl==3.0.7
pandas==1.2.4
passlib==1.7.4
psycopg2==2.8.6
pyasn1==0.4.8
pycparser==2.20
pydantic-sqlalchemy==0.0.8.post1
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.15.0
python-editor==1.0.4
python-http-client==3.3.2
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
redis==3.5.3
requests==2.25.1
rq==1.9.0
rsa==4.7.1
sendgrid==6.6.0
six==1.15.0
sqlalchemy==1.3.23
starkbank-ecdsa==1.1.0
urllib3==1.26.4
uuid==1.30
uvicorn==0.13.3
xlsxwriter==1.4.0
fastapi